+++ src/main/java/in/koreatech/koin/domain/shop/repository/MenuRepository.java
@@ -1,10 +1,10 @@
-package in.koreatech.koin.repository.shop;
+package in.koreatech.koin.domain.shop.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.repository.Repository;
 
-import in.koreatech.koin.domain.shop.Menu;
+import in.koreatech.koin.domain.shop.domain.Menu;
 
 public interface MenuRepository extends Repository<Menu, Long> {
