+++ src/test/java/in/koreatech/koin/acceptance/ShopApiTest.java
@@ -7,12 +7,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 
 import in.koreatech.koin.AcceptanceTest;
-import in.koreatech.koin.domain.shop.Menu;
-import in.koreatech.koin.domain.shop.MenuCategory;
-import in.koreatech.koin.domain.shop.MenuCategoryMap;
-import in.koreatech.koin.domain.shop.MenuImage;
-import in.koreatech.koin.domain.shop.MenuOption;
-import in.koreatech.koin.repository.shop.MenuRepository;
+import in.koreatech.koin.domain.shop.domain.Menu;
+import in.koreatech.koin.domain.shop.domain.MenuCategory;
+import in.koreatech.koin.domain.shop.domain.MenuCategoryMap;
+import in.koreatech.koin.domain.shop.domain.MenuImage;
+import in.koreatech.koin.domain.shop.domain.MenuOption;
+import in.koreatech.koin.domain.shop.repository.MenuRepository;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
